pyproject.toml
[project]
name = "ai-story"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.0",
    "celery==5.5.0b2",
    "channels>=4.0.0",
    "channels-redis>=4.2.1",
    "daphne>=4.2.1",
    "django==3.2.15",
    "django-celery-beat==2.2.1",
    "django-celery-results==2.4.0",
    "django-cors-headers>=4.5.0",
    "django-filter>=23.5",
    "django-redis>=5.4.0",
    "djangorestframework==3.14.0",
    "djangorestframework-jwt>=1.11.0",
    "djangorestframework-simplejwt>=5.3.1",
    "gevent>=25.9.1",
    "jinja2>=3.1.6",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "websocket-client>=1.9.0",
]
